Add scrolling to sprite toolbar
In small screens you will now be able to scroll down if the sprite tool
options is cut off.

By increasing sprite tool options 'bottom' attribute, I changed the height of
the sprite tool options so it doesn't fade behind the sprite helpers.
This also makes it so the scroll bar won't cover up part of the H
Symmetry button.

All items in the sprite tool options decrease in size to make room for
the toolbar automatically.

A little annoying quirk is that decreasing the size of your window will
cause the scroll bar to appear but then it will also decrease the size
of the elements so the scroll bar is unnecessary and it will be a weird
empty scroll bar until you decrease the size of the window further for
it to be necessary again.

In the future it could be nice to center the sprite tool options where
it is, since the left padding looks larger than the right padding. I
think that just means changing the left distance from 58px to 52px.